Rename ProductVersion -> ProductVersionReader and move ClientContext to org.triplea.injection
