+++ game-core/src/main/java/games/strategy/engine/auto/update/EngineVersionCheck.java
@@ -2,7 +2,6 @@ package games.strategy.engine.auto.update;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Splitter;
-import games.strategy.engine.ClientContext;
 import games.strategy.triplea.settings.ClientSetting;
 import games.strategy.triplea.settings.GameSetting;
 import java.time.LocalDate;
@@ -12,6 +11,7 @@ import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
 import lombok.experimental.UtilityClass;
 import lombok.extern.java.Log;
+import org.triplea.injection.ClientContext;
 import org.triplea.live.servers.LiveServersFetcher;
 import org.triplea.swing.EventThreadJOptionPane;
