+++ game-core/src/main/java/games/strategy/engine/framework/map/download/MapDownloadProgressPanel.java
@@ -1,7 +1,6 @@
 package games.strategy.engine.framework.map.download;
 
 import com.google.common.collect.Maps;
-import games.strategy.engine.ClientContext;
 import games.strategy.engine.framework.map.file.system.loader.AvailableGamesFileSystemReader;
 import java.awt.GridLayout;
 import java.util.ArrayList;
@@ -12,6 +11,7 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JProgressBar;
 import javax.swing.SwingUtilities;
+import org.triplea.injection.ClientContext;
 import org.triplea.swing.SwingComponents;
 import org.triplea.swing.jpanel.JPanelBuilder;
