+++ game-core/src/main/java/games/strategy/engine/framework/startup/login/ClientLogin.java
@@ -1,13 +1,13 @@
 package games.strategy.engine.framework.startup.login;
 
 import com.google.common.annotations.VisibleForTesting;
-import games.strategy.engine.ClientContext;
 import games.strategy.net.IConnectionLogin;
 import java.awt.Component;
 import java.util.HashMap;
 import java.util.Map;
 import javax.swing.JOptionPane;
 import javax.swing.JPasswordField;
+import org.triplea.injection.ClientContext;
 import org.triplea.java.Interruptibles;
 import org.triplea.swing.SwingAction;
