+++ game-core/src/main/java/games/strategy/engine/framework/startup/login/ClientLoginValidator.java
@@ -2,7 +2,6 @@ package games.strategy.engine.framework.startup.login;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
-import games.strategy.engine.ClientContext;
 import games.strategy.net.ILoginValidator;
 import games.strategy.net.IServerMessenger;
 import java.net.InetSocketAddress;
@@ -11,6 +10,7 @@ import java.util.Map;
 import java.util.Objects;
 import javax.annotation.Nullable;
 import lombok.Setter;
+import org.triplea.injection.ClientContext;
 import org.triplea.java.Interruptibles;
 import org.triplea.util.Version;
