+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/panels/main/game/selector/GameSelectorPanel.java
@@ -3,7 +3,6 @@ package games.strategy.engine.framework.startup.ui.panels.main.game.selector;
 import static org.triplea.swing.SwingComponents.DialogWithLinksParams;
 import static org.triplea.swing.SwingComponents.DialogWithLinksTypes;
 
-import games.strategy.engine.ClientContext;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.properties.IEditableProperty;
 import games.strategy.engine.data.properties.PropertiesUi;
@@ -38,6 +37,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JPopupMenu;
 import javax.swing.JScrollPane;
+import org.triplea.injection.ClientContext;
 import org.triplea.swing.DialogBuilder;
 import org.triplea.swing.JButtonBuilder;
 import org.triplea.swing.SwingAction;
