+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/posted/game/pbf/test/post/TestPostAction.java
@@ -1,7 +1,6 @@
 package games.strategy.engine.framework.startup.ui.posted.game.pbf.test.post;
 
 import com.google.common.base.Preconditions;
-import games.strategy.engine.ClientContext;
 import games.strategy.engine.posted.game.pbf.NodeBbForumPoster;
 import games.strategy.engine.posted.game.pbf.NodeBbForumPoster.SaveGameParameter;
 import java.awt.Graphics;
@@ -16,6 +15,7 @@ import java.util.logging.Level;
 import javax.imageio.ImageIO;
 import lombok.AllArgsConstructor;
 import lombok.extern.java.Log;
+import org.triplea.injection.ClientContext;
 import org.triplea.java.DateTimeUtil;
 
 @Log
