+++ game-core/src/main/java/games/strategy/engine/random/MartiDiceRoller.java
@@ -3,7 +3,6 @@ package games.strategy.engine.random;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
-import games.strategy.engine.ClientContext;
 import games.strategy.engine.framework.system.HttpProxy;
 import java.io.IOException;
 import java.net.URI;
@@ -33,6 +32,7 @@ import org.apache.http.impl.client.LaxRedirectStrategy;
 import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.protocol.HttpContext;
 import org.apache.http.util.EntityUtils;
+import org.triplea.injection.ClientContext;
 
 /** A pbem dice roller that reads its configuration from a properties file. */
 @Builder
