+++ game-core/src/main/java/org/triplea/config/product/ProductVersionReader.java
@@ -9,15 +9,15 @@ import org.triplea.util.Version;
  * Provides access to the product configuration. The product configuration applies to all components
  * of the TripleA application suite (e.g. game client, lobby server, etc.).
  */
-public final class ProductConfiguration {
+public final class ProductVersionReader {
   private final PropertyReader propertyReader;
 
-  public ProductConfiguration() {
+  public ProductVersionReader() {
     this(new ResourcePropertyReader("META-INF/triplea/product.properties"));
   }
 
   @VisibleForTesting
-  ProductConfiguration(final PropertyReader propertyReader) {
+  ProductVersionReader(final PropertyReader propertyReader) {
     this.propertyReader = propertyReader;
   }
