+++ game-core/src/main/java/org/triplea/debug/error/reporting/formatting/ErrorReportBodyFormatter.java
@@ -1,7 +1,6 @@
 package org.triplea.debug.error.reporting.formatting;
 
 import com.google.common.base.Strings;
-import games.strategy.engine.ClientContext;
 import games.strategy.engine.framework.system.SystemProperties;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintWriter;
@@ -13,6 +12,7 @@ import javax.annotation.Nullable;
 import lombok.experimental.UtilityClass;
 import org.triplea.debug.ExceptionDetails;
 import org.triplea.debug.LoggerRecord;
+import org.triplea.injection.ClientContext;
 
 /**
  * Based on a LogRecord, creates the body details of an error report. The body is the 'main' part of
