+++ game-core/src/main/java/org/triplea/injection/ClientContext.java
@@ -1,7 +1,7 @@
-package games.strategy.engine;
+package org.triplea.injection;
 
 import games.strategy.engine.framework.map.download.DownloadCoordinator;
-import org.triplea.config.product.ProductConfiguration;
+import org.triplea.config.product.ProductVersionReader;
 import org.triplea.util.Version;
 
 /**
@@ -38,7 +38,7 @@ import org.triplea.util.Version;
 public final class ClientContext {
   private static final ClientContext instance = new ClientContext();
 
-  private final ProductConfiguration productConfiguration = new ProductConfiguration();
+  private final ProductVersionReader productVersionReader = new ProductVersionReader();
   private final DownloadCoordinator downloadCoordinator = new DownloadCoordinator();
 
   private ClientContext() {}
@@ -48,6 +48,6 @@ public final class ClientContext {
   }
 
   public static Version engineVersion() {
-    return instance.productConfiguration.getVersion();
+    return instance.productVersionReader.getVersion();
   }
 }
