+++ game-core/src/test/java/games/strategy/engine/ClientContextIntegrationTest.java
@@ -10,6 +10,7 @@ import games.strategy.engine.framework.map.listing.MapListingFetcher;
 import games.strategy.triplea.settings.AbstractClientSettingTestCase;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.triplea.injection.ClientContext;
 import org.triplea.test.common.Integration;
 
 @Integration
